Changed value types of balance and bets from double to float, added a Bet dataclass

+++ src/main/kotlin/casino/core/payment/IPaymentService.kt
@@ -3,5 +3,5 @@ package main.kotlin.casino.core.payment
 import main.kotlin.casino.core.domain.Player
 
 interface IPaymentService {
-    fun addFunds(player: Player, amount: Double)
+    fun addFunds(player: Player, amount: Float)
 }
